add xbmc_codec_descriptor.hpp to reduce code duplication
