+++ addons/pvr.hts/src/HTSPTypes.h
@@ -28,42 +28,7 @@
 #include "platform/util/StdString.h"
 #include "libXBMC_codec.h"
 #include "client.h"
-
-class CodecDescriptor
-{
-public:
-  CodecDescriptor(void)
-  {
-    m_codec.codec_id   = XBMC_INVALID_CODEC_ID;
-    m_codec.codec_type = XBMC_CODEC_TYPE_UNKNOWN;
-  }
-
-  CodecDescriptor(xbmc_codec_t codec, const char* name) :
-    m_codec(codec),
-    m_strName(name) {}
-  virtual ~CodecDescriptor(void) {}
-
-  const std::string& Name(void) const  { return m_strName; }
-  xbmc_codec_t Codec(void) const { return m_codec; }
-
-  static CodecDescriptor GetCodecByName(const char* strCodecName)
-  {
-    CodecDescriptor retVal;
-    // some of Tvheadend's and VDR's codec names don't match ffmpeg's, so translate them to something ffmpeg understands
-    if (!strcmp(strCodecName, "MPEG2AUDIO"))
-      retVal = CodecDescriptor(CODEC->GetCodecByName("MP2"), strCodecName);
-    else if (!strcmp(strCodecName, "MPEGTS"))
-      retVal = CodecDescriptor(CODEC->GetCodecByName("MPEG2VIDEO"), strCodecName);
-    else
-      retVal = CodecDescriptor(CODEC->GetCodecByName(strCodecName), strCodecName);
-
-    return retVal;
-  }
-
-private:
-  xbmc_codec_t m_codec;
-  std::string  m_strName;
-};
+#include "xbmc_codec_descriptor.hpp"
 
 typedef std::vector<CodecDescriptor> CodecVector;
